// Anomologita.Mobile/types/auth.types.ts
import { Post } from "./post.types";
export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  accessToken: string;
  accessTokenExpiry: string;
  refreshToken: string;
  refreshTokenExpiry: string;
  userId: string;
  username: string;
  isAdmin: boolean;
  isStudent: boolean;
  universityId: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  confirmPassword: string;
}

export interface RegisterResponse {
    message: string;
}

export interface RefreshRequest {
    refreshToken: string;
}

export interface RefreshResponse {
    accessToken: string;
    accessTokenExpiry: string;
    refreshToken: string;
    refreshTokenExpiry: string;
}

export interface FetchUserResponse {
    id: string;
    username: string;
    email: string;
    universityId: boolean;
    universityName: string;
    universityShortName: string;
    createdAt: string;
    posts: Post[];
}